Add a copy script to copy over images from src to dist and update the dev script to use the copy script.
Images need to be moved to /dist in order to be statically served. Parcel does not seem to have this functionality out of the box, so the solution is to make a basic copy script. The drawback is there is no watching, so the script will need to be re-run to copy any images that are added during development.